Added new Rule Category, PVP. This will be used for specific rules related to PVP that doesn't always fit in the other categories.
Added rule to enable/disable when a inhabited race of a city is player killed, the killer's faction with that city is lowered. The actual faction is controlled by quest files player.lua in each city.

Rule
-------------------
RULE_BOOL(PVP, FactionCityKilling, false, "When enabled player killing the inhabitants of a city lowers the killers faction.")